+++ src/main/java/com/vape/data/streaming/model/RMSDataPointModel.java
@@ -5,8 +5,6 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 @Document(collection="rms")
@@ -15,5 +13,5 @@ public class RMSDataPointModel {
     private String id;
     private String sensorDataPointId;
     private String timestamp;
-    private BigDecimal data;
+    private Double data;
 }
